Create bridge finder instead of modifiying bridge paring.
