+++ WinHue3/Functions/BridgePairing/BridgePairingViewModel.cs
@@ -29,7 +29,7 @@ namespace WinHue3.Functions.BridgePairing
         private readonly DispatcherTimer _pairTimer = new DispatcherTimer();
         private string _scanButtonText;
         private ObservableCollection<Bridge> _listBridges;
-        private BackgroundWorker _bgw;
+
 
         public BridgePairingViewModel()
         {
@@ -42,79 +42,7 @@ namespace WinHue3.Functions.BridgePairing
             Hue.OnBridgeDetectionFailed += Hue_OnBridgeDetectionFailed;
             _scanButtonText = GUI.BridgeDetectionPairing_Scan;
             _listBridges = new ObservableCollection<Bridge>();
-            _bgw = new BackgroundWorker();
-            _bgw.DoWork += _bgw_DoWork;
-            _bgw.RunWorkerCompleted += _bgw_RunWorkerCompleted;
-        }
 
-        private void _bgw_RunWorkerCompleted(object sender, RunWorkerCompletedEventArgs e)
-        {
-            
-        }
-
-        private void _bgw_DoWork(object sender, DoWorkEventArgs e)
-        {
-            string mac = e.Argument.ToString();
-            IPAddress ip = IPAddress.Parse(GetLocalIPAddress());
-            byte[] ipArray = ip.GetAddressBytes();
-            byte currentip = ipArray[3];
-
-            for (byte x = 2; x <= 254; x++)
-            {
-                if (_bgw.CancellationPending)
-                {
-                    log.Info("Bridge lookup cancelled.");
-                    e.Cancel = true;
-                    break;
-                }
-
-                _bgw.ReportProgress(0, x);
-                if (x == currentip) continue;
-                ipArray[3] = x;
-
-                Comm.Timeout = 50;
-                if (_bgw.CancellationPending) break;
-
-                CommResult comres = Comm.SendRequest(new Uri($@"http://{new IPAddress(ipArray)}/api/config"), WebRequestType.Get);
-                Philips_Hue.BridgeObject.BridgeObjects.BridgeSettings desc = new Philips_Hue.BridgeObject.BridgeObjects.BridgeSettings();
-
-                switch (comres.Status)
-                {
-                    case WebExceptionStatus.Success:
-                        desc = Serializer.DeserializeToObject<Philips_Hue.BridgeObject.BridgeObjects.BridgeSettings>(comres.Data); // try to deserialize the received message.
-                        if (desc == null) continue; // if the deserialisation didn't work it means this is not a bridge continue with next ip.
-                        if (desc.mac == mac)
-                        {
-                            e.Result = new IPAddress(ipArray);
-                            break;
-                        }
-
-                        break;
-                    case WebExceptionStatus.Timeout:
-                        // IP DOES NOT RESPOND
-                        break;
-                    default:
-
-                        break;
-                }
-            }
-        }
-
-        /// <summary>
-        /// Get the current IP of the computer to scan current range for bridge.
-        /// </summary>
-        /// <returns></returns>
-        private static string GetLocalIPAddress()
-        {
-            var host = Dns.GetHostEntry(Dns.GetHostName());
-            foreach (var ip in host.AddressList)
-            {
-                if (ip.AddressFamily == AddressFamily.InterNetwork)
-                {
-                    return ip.ToString();
-                }
-            }
-            throw new Exception("Local IP Address Not Found!");
         }
 
         public ObservableCollection<Bridge> ListBridges
@@ -288,11 +216,6 @@ namespace WinHue3.Functions.BridgePairing
             SelectedBridge.IsDefault = true;
             RaisePropertyChanged("DefaultSet");
             log.Info($@"Settings default bridge to {_selectedBridge}");
-
-        }
-
-        public void FindBridge(Bridge br)
-        {
 
         }
